requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
pydantic==2.10.1
python-dotenv==1.0.1

# OpenAI API client (for custom base URI support)
openai==1.57.2

# Async support
httpx==0.27.2
aiofiles==24.1.0

# Type hints
typing-extensions==4.12.2

# Optional: for enhanced JSON handling
orjson==3.10.12
